pls: move server bits to separete file
In preparation to handle multiple Perl servers, this commit move the bits
related to the Perl::LanguageServer server to its specific typescript file,
allowing further modularization for the next server we want support.
